+++ test/test_vmec.py
@@ -1,0 +1,6 @@
+from simsopt.mhd import Vmec
+
+# def test_vmec():
+
+# equil = Vmec("data/w7x/wout_d23p4_tm.nc")
+# import pdb;pdb.set_trace()